Merged PR 446715: Fix DSM arguments to shared page not being stored
Make the INPT buffer be the exact same size as MBUF, so ASL does not do type conversion and instead copies byte-for-byte the contents of INPT into MBUF, so the PMEM VDEV can see the arguments sent by the guest. Unsure if this is the best way to do this, but it works.