Merged PR 3153266: Fix VmbfsDxe passing a guid, not pointer, to UninstallMultipleProtocolInterfaces
On X64, via the ABI 128 bit arguments are passed via pointer, so the compiler actually passes a pointer to the GUID which gets spilled onto the stack, which is what the function expects. Basically, we get lucky.

On AArch64, the ABI is different and instead the 2nd argument register has a portion of the guid, which is not a valid address, leading to an access violation.

TESTED: ARM64 VMBFS boot. X64 UVM VMBFS boot.

Related work items: #21228154