Merged PR 6511987: Common fixes for isolated platforms with no paravisor
These are the common fixes that are needed to boot an SNP partition with no paravisor. The remaining changes can be found at the `feature/snp-enlightened` branch which will not be part of this PR as we currently have no plans to run SNP enlightened UEFI in this mode for CVM. The changes to implement a #HV handler would require invasive changes to CpuDxe and EfiHvDxe, and will continue to stay in the feature branch.

TESTED: Boots to frontpage when running under SNP with `feature/snp-enlightened`. Normal UEFI boots to frontpage.

Related work items: #28605969