Merged PR 8110436: Fix a bug where the wrong page count was being passed to page acceptance routines
Fix a bug where the wrong page count was being passed to page acceptance routines.
As a result of this bug, pages were being marked as exclusive but not accepted. This would lead to later #VEs within the normal OS since as far as it knew, the pages were private and accepted.

Related work items: #42172450