Merged PR 8201222: Extend the MP wake mailbox protocol's page tables to identity map in all memory < 2MB
The specification only requires the wakeup vector to be identity mapped, but this is challenging because it would require building page tables per AP from the mailbox (firmware) code. While pursuing a specification change, take advantage of the fact that Linux specifies a below 1MB page (on the lines of INIT-SIPI) as the wakeup vector and identity map all memory in the low 2MB region into the page tables constructed for the mailbox.

Related work items: #40035798