Merged PR 8616802: Fix the waitForEvent calls so that we can wait at any TPL
The check for TPL  when waiting for events was brought back in basecore with the following change:

https://microsoft.visualstudio.com/DefaultCollection/OS/_git/hyperv.uefi.mu_basecore/commit/202a9bb8c0df5d286b9bfad8539b28140f628692?refName=refs%2Fheads%2Frelease%2F202208&path=%2FMdeModulePkg%2FCore%2FDxe%2FEvent%2FEvent.c&_a=compare

Because of this, our waits started failing with the wrong TPL error.

This UEFI change uses a different protocol for waiting for events which does not have the strict TPL enforcement.

Testing:
- Ran the WDS boot scenario described in the bug.
- Ran the Uefi.VpciNvmeBoot.Tests.ps1 test
- Run a VM with IMC enabled.
- VmBoot.Tests.ps1 for vmbfs boot.

Related work items: #43534654