Merged PR 10409763: Support 64k Runtime Page Allocation Granularity on ARM64
This PR adds ARM64 64k Runtime Page Allocation Granularity support. It does not pull in the mu_basecore changes required for this yet, as they are still in PR form, but this PR will be updated to include them.

This PR splits the ARM64 FDF into two FVs, a SEC/PEI FV that is rebased by the basetools (as it contains XIP code) and a DXE FV that is not rebased by the basetools (as it does not contain any XIP code). This is required as this changes DXE_RUNTIME_DRIVERs for ARM64 to have 64k section alignment (which is required per UEFI spec) and the basetools do not support rebasing them. This new FV is also published as an FV HOB for DxeCore to consume. DxeCore itself is in the SEC/PEI FV so that DxeIpl can find it. This change has the additional benefit that PeiCore no longer has to search through all of the DXE drivers to determine if it needs to load them, in theory this should speed up booting.

In addition, this disables page and pool guards for ARM64 for EfiRuntimeServicesCode, EfiRuntimeServicesData, EfiACPIMemoryNVS, and EfiReservedMemoryType, as the heap guard system does not support guarding 64k aligned regions.

Testing:
- Boot X64 VM.
- Hibernate X64 VM and resume with the new firmware.
- Boot ARM64 VM

Related work items: #40041681, #49021885