Merged PR 10597080: Remove Exdi driver
Remove Exdi driver

This Exdi driver is relying on the windbg heuristic to find a named and structured list which mimics NT structures, we should use the existing UEFI structures in DxeCore. This code uses the UEFI structures to transpose manually to discover the module at each header address.

DxeCore keeps a list of loaded modules called mDebugInfoTableHeader. The correct way is to use the UefiExt debug extension to parse this and load the modules symbols using !uefiext.findall

Related work items: #48874140