Merged PR 10817701: Separate rdrand from Shared Crypto (2023.2.15)
This update removes rdrand assumptions from the shared crypto binary
including implicit assumptions for its support for randomizing stack
cookie values.

RNG support is now consumed through the newly introduced RNG PPI.
This change also adds RngPei to the platform build to produce the
PPI so it can be consumed by the Shared Crypto PEI binary.

Currently, in MsvmPkg, the RNG PPI is backed by the NULL instance
of RngLib. So, it is not actually functional. MsvmPkg should
update MsvmPkg/Library/RngLib/RngLib.inf to support PEI and then
link that library instance to RngPei instead of the NULL instance.

This will use pre-existing logic in the library to check for rdrand
support via cpuid to determine whether to use the instruction or
defer to host emulation.

Related work items: #50829064