Merged PR 11232433: VmBus: support confidential control plane on CVM
This change updates the UEFI VmBus client with support for the confidential control plane, which lets it connect to VmBus services offered by the paravisor of a hardware isolated VM by using regular hypercalls and encrypted memory. It is equivalent to !11022303 which made the same changes for the Windows guest VmBus driver.

This change will be followed by an additional change to allow the use of encrypted ring buffers.

This change updates the protocol used by the UEFI client the Copper VmBus protocol. This will be needed for the encrypted ring buffer support. I believe this is change since the UEFI firmware is provided by the host or paravisor, so it will never run on a host that doesn't support this protocol version.

----
#### AI description  (iteration 1)
#### PR Classification
New feature: support for confidential control plane on CVM in VmBus.

#### PR Summary
This pull request adds support for a confidential control plane on CVM in VmBus, addressing the linked work item for UEFI support.
- `MsvmPkg/EfiHvDxe/EfiHv.c`: Added functions to enable and disable synthetic interrupt controller components, and updated connection logic to support confidential control plane.
- `MsvmPkg/VmbusDxe/ChannelMessages.h`: Introduced new message structures and feature flags for confidential channels and updated existing message structures to support new features.
- `MsvmPkg/VmbusDxe/VmbusRoot.c`: Added a function to handle SINT connection with support for confidential channels.

Related work items: #53097226