Merged PR 11331849: VmBus: support encrypted ring buffers on CVM.
This change enables the UEFI VmBus client to use confidential channels, which are paravisor-offered channels on a hardware isolated VM that can use an encrypted ring buffer, not requiring the memory to be host visible.

This is the UEFI version of !11170874, which made the same change for the Windows VmBus client.

Because confidential channels require the Copper protocol, this change also implements version negotiation so UEFI can use Copper when available and fall back when it's not.

This was tested using a TDX-isolated Underhill VM, using a VTL2-offered storage channel backed by the NVMe emulator.

Related work items: #53458154