Merged PR 11410015: VmBus: Fix version response size check.
!11331849 introduced an error in the way the size of the VersionResponse message was validated. This message has a larger size when using the Copper protocol, and the change incorrectly assumed that the host would respond with that size if Copper was requested, but if the host doesn't support Copper that's not the case.

This change fixes the size check to only check for the larger message if the host responds that it supports the Copper version.

This was tested by running a version 11 VM (supporting Copper) and a version 10.5 VM (pre-Copper).

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to ensure compatibility with older VM protocols.

#### PR Summary
This pull request addresses a bug in the version response size check to support older protocols from pre-Copper VM versions.
- `MsvmPkg/VmbusDxe/VmbusRoot.c`: Removed conditional size check and enforced a minimum size check for version response. Added a specific size check for Copper version responses.

Related work items: #53814096