Merged PR 11658431: VmBus: support confidential external memory on CVM.
This change adds VmBus client support for an additional offer flag that indicates memory other than the ring buffer must also be encrypted. If this flag is present, any additional GPADLs are not made host visible, and bounce buffering is disabled for GPA direct packets.

Support for this offer flag is indicated by the same feature flag that was already used for encrypted ring buffers. Underhill can set these flags when appropriate when running on a hardware-isolated VM.

This is the UEFI equivalent of !11612955

The change was tested by using a TDX-isolated VM with the NVMe emulator, using a flag added to Underhill to force the use confidential external memory (as there are currently no devices that actually support this). The OS was booted off this device, testing whether UEFI could read it during boot with this flag set. I have also validated the new firmware works correctly without the new flag, and on non-confidential VMs.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds support for confidential external memory in CVM for VmBus.
- `MsvmPkg/VmbusDxe/VmbusChannel.c`: Added handling for confidential external memory and ring buffer flags.
- `MsvmPkg/EmclDxe/Emcl.c`: Updated GPADL preparation to use ring buffer flag.
- `MsvmPkg/Include/Protocol/Vmbus.h`: Introduced new flags for confidential external memory and ring buffer.
- `MsvmPkg/VmbusDxe/VmbusP.h`: Modified channel context to distinguish between confidential ring buffer and external memory.
- `MsvmPkg/VmbusDxe/ChannelMessages.h`: Defined new offer flags for confidential ring buffer and external memory.

Related work items: #54586579