Merged PR 12451610: Fix nuget name clash for legacy debugger job
CI for main failed because DEBUG builds with and without the legacy debugger shared the same nuget package name. This change conditionally checks for that and skips the nuget publishing on legacy

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a name clash issue for the legacy debugger job in the NuGet package.
- Changes in `/MsvmPkg/OneBranch/Build-And-Publish.yaml` to conditionally append `Build.Variant` to the package ID for legacy builds.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #56508688